Modified the BardDomainModel plugin
- Configured the bard-domain-model plugin to load after the database-migration plugin
- set the migrationResourceAccessor to be relative to the bard-domain-model plugin install dir
- update the changelogs to use the configured resourceAccessor

This allows a top level grails app that includes the bard-domain-model plugin to get the database-migraion plugin
as a transitive dependency and all the the top level app to execute the database-migration scripts.  The scripts like
dbm-update will look in the bard-domain-model plugin dir for the changelogs instead of the top level app.