updated some collections to use the lazy : false flag vs fetch : 'join'
seemed like with join the lists where not being consistently returned sometimes, items were missing.

I try and reporduce this in an isolated example and see if it's still and issue with the latest grails.
Switching to lazy: false seemed to consistently give the correct number of items in the lists.