Modified the repository config to not inherit repositories from plugins
We want all artifacts cached in our artifactory server, if a plugin needs an artifact from a repo we currently do not proxy
we should update our artifactory instance and add the proxy and work it into the resolution order in the right place.

Along with these changes, I also tweaked the config of artifactory to try and eleminate situations were remote repos were timing out.