Experiment with extending TransactionAwareDataSourceProxy and overwriting dataSource
The grails datasource plugin proxies datasources in a TransactionAwareDataSourceProxy, experimenting with
having the DomainPlugin supply a custom implementation where we execute a store procedure as connections are requested
and set the current user based on the user that authenticated with SpringSecurity.

Additionally, added a function to clear the user from the context in the database and set the testOnReturn or the commons dbcp to true and
specified a validationQuery that clears out the username from the session.

Commons dbcp didn't have great hooks to intercept connections being checked out and returned to the pool.  Not sure if this is the cleanest way,
but it does get us what we need in the short term.

Looked a changing the connection pool, but that was a bit more involved than I had hoped.