fix for https://www.pivotaltracker.com/story/show/50915009
When valueNum, valueMin or valueMax have values of zero the validation logic isn't doing the right thing

updated the tests and the code to have this work properly. Now it's possible to save contextItem with a 0.0 value in any of these number fields and either succeed or get reasonable validation messages. Previously, the zero value resulted in the wrong validation branch being executed.