Fixed problem where cloning would not show edit buttons after clone.
This was because it was rendering the page, but this code path didn't have
the editable flag added to the model used by the template.   Changed to use
redirect instead so that the template was only used from one place.

This in turn exposed a problem where the show() action was using the flash scope when it should not have been.  Fixed that as well so that the error message shows up when assay cannot be cloned.

https://www.pivotaltracker.com/story/show/54059075