fix for [#66340356] move logic for assayStatus validation check to service
adding a validation that pulled in contexts and contexts items introduces hibernate exceptions in some cases, the options were to either load the dependent objects with a new hibernate session or not integrate these directly in the GORM validation.  Decided to not directly iterate this with the GORM validation and call this check verifying no context items had null values to the service method that was already handling the update of the assay status.